requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
captum==0.7.0
numpy==2.0.1
pandas==2.2.3
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.6.0
scipy==1.13.1
pillow==11.0.0
tqdm==4.67.1
fsspec==2024.10.0
joblib==1.4.2
python-dateutil==2.9.0post0
typing-extensions==4.11.0
requests==2.32.3
urllib3==2.2.3
